#14: rename to util module darkModeSettingUtils
